requirements.txt
python-dotenv
plexapi
openai
google-api-python-client
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
markdown2
